+++ src/types/index.ts
@@ -7,6 +7,10 @@ export enum Amm {
     MILADY = "milady",
     PPG = "ppg",
     CDB = "cdb",
+    CAP = "cap",
+    SPROTO = "sproto",
+    SOFA = "sofa",
+    DEGODS = "degods",
 }
 
 export enum Side {